Add local changes on georgebarnick.xyz
Kind of messy right now, but disabled a lot of global stuff and adjusted
some config for the new server. Any instance of "georgebarnick" or
"georgebarnick.xyz" will need to be changed to "brickimedia" and
"brickimedia.org" once the DNS is reconfigured. Some global stuff will
need to be redone as well, such as images being hosted on
www.brickimedia.org and being usable on greatballcontraption.com, global
avatars, shared database, etc.