Undeploy SiteMatrix
It was always a WMF-specific extension, with many interface messages explicitly mentioning Wikimedia, for example. Given the recent decision to close down and merge all of our other wikis (save for GBC) into Brickipedia, having SiteMatrix enabled no longer makes any sense.

I'm not sure if $wgLocalDatabases and $wgConf->wikis definitions are safe to remove or not, so I'm leaving those in for the time being.